fixing the folder order + added reset to clearance in first task
