+++ HW2/arch/x86/entry/syscalls/syscall_64.tbl
@@ -341,6 +341,9 @@
 332	common	statx			sys_statx
 333	common  hello			sys_hello
 334	common	set_sec			sys_set_sec
+335	common	get_sec			sys_get_sec
+336	common	check_sec		sys_check_sec
+337	common	set_sec_branch			sys_set_sec_branch
 
 #
 # x32-specific system call numbers start at 512 to avoid cache impact
