+++ HW2/kernel/syscalls.h
@@ -942,4 +942,7 @@ asmlinkage long sys_statx(int dfd, const char __user *path, unsigned flags,
 			  unsigned mask, struct statx __user *buffer);
 asmlinkage long sys_hello(void);
 asmlinkage long sys_set_sec(int sword, int midnight, int clamp);
+asmlinkage long sys_get_sec(char clr);
+asmlinkage long sys_check_sec(pid_t pid, char clr);
+asmlinkage long sys_set_sec_branch(int height, char clr);
 #endif
